+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Rename `src/display/styles/help.rs` to `cli_help.rs` for clarity
 - Rename `src/display/styles/text.rs` to `value.rs` for clarity
 
+### Fixed
+- Delayed summary on large directories
+
 ## [0.1.21] - 2026-02-09
 ### Added
 - `apkm` and `xapk` extensions to the Android icon